pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sin-photonic-mcp-agent"
version = "0.1.0"
description = "Agentic AI for SiN photonic waveguide design using MCP-served physics tools (modesolverpy, SAX, gdsfactory)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "A Taylor"}]
dependencies = [
    "strands-agents>=0.3.0",
    "strands-agents-tools>=0.2.0",
    "bedrock-agentcore>=0.1.0",
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "datasets>=3.0.0",
    "aws-opentelemetry-distro>=0.10.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "joblib>=1.4.0",
    "pydantic>=2.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
physics = [
    "modesolverpy>=0.5.0",
    "sax>=0.14.0",
    "jax>=0.4.30",
    "jaxlib>=0.4.30",
    "gdsfactory>=8.0.0",
    "gdstk>=0.9.0",
]
storage = [
    "pyarrow>=17.0.0",
    "s3fs>=2024.6.0",
]
mcp = [
    "fastmcp>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
